refactoring, update data entry form
