+++ app/src/main/AndroidManifest.xml
@@ -21,7 +21,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".ProductDetails"
+        <activity android:name=".ui.productdetails.ProductDetails"
             android:parentActivityName=".CartActivity"/>
         <activity
             android:name=".MainActivity"
