+++ app/src/main/java/com/krisbunda/gamesmart/ProductsAdapter.kt
@@ -7,9 +7,10 @@ import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
-import com.krisbunda.gamesmart.ProductsAdapter.ViewHolder
 import com.krisbunda.gamesmart.model.Product
+import com.krisbunda.gamesmart.ui.productdetails.ProductDetails
 import com.squareup.picasso.Picasso
+import kotlin.math.roundToInt
 
 
 //pass Products Adapter from CartActivity
@@ -18,11 +19,11 @@ class ProductsAdapter(private val products: List<Product>) : RecyclerView.Adapte
     //bind objects to product row layout
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
         val product = products[position]
-        Picasso.get().load(product.photoUrl).into(holder.image)
         holder.title.text = product.title
+        Picasso.get().load(product.photoUrl).into(holder.image)
         holder.descprod.text = product.descProd
-        holder.pricedol.text = product.price.toString()
-        holder.pricepts.text = product.points.toString()
+        holder.pricedol.text = "$${product.price?.roundToInt()}"
+        holder.pricepts.text = "${product.points}pts"
 
     }
 
@@ -34,9 +35,9 @@ class ProductsAdapter(private val products: List<Product>) : RecyclerView.Adapte
         view.setOnClickListener {
             val intent = Intent(parent.context, ProductDetails::class.java)
             intent.putExtra("title",products[holder.adapterPosition].title)
-            intent.putExtra("proddesc",products[holder.adapterPosition].descProd)
             intent.putExtra("photoaddress",products[holder.adapterPosition].photoUrl)
-            intent.putExtra("pricedol",products[holder.adapterPosition].price.toString())
+            intent.putExtra("proddesc",products[holder.adapterPosition].descProd)
+            intent.putExtra("pricedol",products[holder.adapterPosition].price?.roundToInt().toString())
             intent.putExtra("pricepts",products[holder.adapterPosition].points.toString())
 
             parent.context.startActivity(intent)
@@ -48,8 +49,8 @@ class ProductsAdapter(private val products: List<Product>) : RecyclerView.Adapte
     override fun getItemCount() = products.size
 
     class ViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView){
-        val image: ImageView = itemView.findViewById(R.id.photo)
         val title: TextView = itemView.findViewById(R.id.title)
+        val image: ImageView = itemView.findViewById(R.id.photo)
         val descprod: TextView = itemView.findViewById(R.id.description)
         val pricedol: TextView = itemView.findViewById(R.id.price)
         val pricepts: TextView = itemView.findViewById(R.id.points)
