+++ app/src/main/java/com/krisbunda/gamesmart/ui/prodcategories/MakerKitsFragment.kt
@@ -1,4 +1,4 @@
-package com.krisbunda.gamesmart.model
+package com.krisbunda.gamesmart.ui.prodcategories
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,12 +7,12 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.krisbunda.gamesmart.R
 
-class PuzzlesFragment : Fragment() {
+class MakerKitsFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragment_cat_puzzles, container, false)
+        return inflater.inflate(R.layout.fragment_cat_makerkits, container, false)
     }
 }
