+++ app/src/main/java/com/krisbunda/gamesmart/ui/productdetails/ProductDetails.kt
@@ -1,12 +1,14 @@
-package com.krisbunda.gamesmart
+package com.krisbunda.gamesmart.ui.productdetails
 
 import android.content.DialogInterface
 import android.os.Bundle
 import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
+import com.krisbunda.gamesmart.R
 import com.squareup.picasso.Picasso
 
 import kotlinx.android.synthetic.main.product_details.*
+import kotlinx.android.synthetic.main.product_details.photo
 
 class ProductDetails : AppCompatActivity() {
 
@@ -20,12 +22,11 @@ class ProductDetails : AppCompatActivity() {
         val price = intent.getStringExtra("pricedol")
         val points = intent.getStringExtra("pricepts")
 
-
         product_name.text = title
         product_desc.text = productdesc
         Picasso.get().load(photoUrl).into(photo)
-        price_dol.text = price
-        price_pts.text = points
+        price_dol.text = "$${price}"
+        price_pts.text = "${points}pts"
 
 
         availability.setOnClickListener {
