+++ app/src/main/java/com/krisbunda/gamesmart/ui/shopadmin/ShopAdminFragment.kt
@@ -1,4 +1,4 @@
-package com.krisbunda.gamesmart.model
+package com.krisbunda.gamesmart.ui.shopadmin
 
 import android.os.Bundle
 import android.util.Log.d
@@ -10,8 +10,6 @@ import androidx.room.Room
 import com.krisbunda.gamesmart.R
 import com.krisbunda.gamesmart.database.AppDatabase
 import com.krisbunda.gamesmart.database.ProductData
-import kotlinx.android.synthetic.main.fragment_shop.*
-import kotlinx.android.synthetic.main.fragment_shop.view.*
 import kotlinx.android.synthetic.main.fragment_shop_admin.*
 import org.jetbrains.anko.doAsync
 import org.jetbrains.anko.uiThread
@@ -35,6 +33,9 @@ class ShopAdminFragment : Fragment() {
 
         btnProdUpdate.setOnClickListener {
             val title = inputProdTitle.text
+            val descProd = inputProdDesc.text
+            val price = inputProdPrice.text
+            val points = inputProdPoints.text
 
             d("dbtest", "Prod Update btn pressed, with items: $title")
 
@@ -43,18 +44,18 @@ class ShopAdminFragment : Fragment() {
 
                 val db: AppDatabase = Room.databaseBuilder(
                     activity!!.applicationContext,
-                    AppDatabase::class.java, "ProdDataDb"
+                    AppDatabase::class.java, "ProductData"
                 ).build()
 
 
                 db.productDao().insertAll(ProductData(
                     null,
                     title.toString(),
-                    //"https://via.placeholder.com/300/BB86FC/FFFFFF/?text=GameSmart",
-                    //it.photoUrl,
-                    "ipsum in voluptate fugiat irure pariatur mollit non deserunt reprehenderit dolore id officia sit",
-                    65.00,
-                    65000
+                    "https://via.placeholder.com/300/BB86FC/FFFFFF/?text=GameSmart",
+                    descProd.toString(),
+                    //"ipsum in voluptate fugiat irure pariatur mollit non deserunt reprehenderit dolore id officia sit",
+                    price.toString().toDoubleOrNull(),
+                    points.toString().toInt()
                 ))
                 val products = db.productDao().getAll()
