+++ app/src/main/res/layout/fragment_shop.xml
@@ -16,6 +16,7 @@
         android:layout_marginEnd="8dp"
         android:layout_marginRight="8dp"
         android:hint="Search"
+        android:backgroundTint="@color/primaryTextColor"
         android:textColor="@color/primaryTextColor"
         android:textColorHint="@color/secondaryColor"
         app:boxStrokeColor="@color/secondaryColor"
